requirements.txt
google-api-python-client==2.1.0
flask==1.1.2
slackclient==2.9.3
gspread==3.7.0
emoji==1.2.0
num2word~=1.0.1
mock~=4.0.3
pytest~=6.2.3